+++ src/middlewares/index.js
@@ -1,0 +1,7 @@
+const getWebClientMiddleware = require('./getWebClientMiddleware');
+const getChannelMembersMiddleware = require('./getChannelMembersMiddleware');
+
+module.exports = {
+  getWebClientMiddleware,
+  getChannelMembersMiddleware
+}
